Fix issue #1289

+++ ibm/resource_ibm_is_lb_listener.go
@@ -115,7 +115,7 @@ func resourceIBMISLBListenerCreate(d *schema.ResourceData, meta interface{}) err
 	var connLimit int64
 
 	if limit, ok := d.GetOk(isLBListenerConnectionLimit); ok {
-		connLimit = limit.(int64)
+		connLimit = int64(limit.(int))
 	}
 
 	client := lbaas.NewLoadBalancerClient(sess)
